Fix vault error message parsing
The make file doesn't use bash but sh, which parses env vars differently. To
escape the envvar VAULT_AUTH_GITHUB_TOKEN in sh, we need to use single quotes
for the string, and a double dollar sign before the name of the envvar. This
will stop the value of the env var getting displayed in the terminal after the
user has copy/pasted it.