Add flag for disabling configuration of git hooks
In most cases it's sensible for `n-gage` to automatically configure
the git hooks that are run by Husky, however sometimes this behaviour
isn't desirable e.g. if you have `n-gage` configured in a monorepo,
where the git hooks are needed at the root of the project, but not
in the directory for a package where `n-gage` is used to take care
of things like installing dependencies and generating an `.env` file.