wait for approval before starting renovate PR builds
renovate PRs for common dependencies often fill up our circleci queue. a paused job waiting for approval doesn't count towards concurrent jobs, and it's straightforward to start this build if we need to test the PR on CI.

because this branch name uses the same format as renovate, it should match the rule so we can test this.