Updated README with WSL troubleshooting
As an WSL user, I was getting `env: 'Files': Permission denied` error for each `make` command I ran in few different FT repos. I looked up into n-gage source and I was able to locate the reason. It is not caused by n-gage or make, but by having Windows paths with spaces added to Linux path. Got this solution https://github.com/Microsoft/WSL/issues/1766#issuecomment-923959283 after googling and described the problem and the workaround into this PR. I have applied on my system and it solved the problem. Suggesting to add it to the README.