Don't run npm update in Heroku postbuild task
heroku-postbuild is a task that is typically called via a
`heroku-postbuild` script in the project's `package.json`, which will be
called during a Heroku build after the main build script has run. One
command in the task is `npm update`. Since npm 8.3.2, this command will
now automatically prune dev dependencies if running in a production Node
environment (it's unclear if this was a conscious change.) This means
that subsequent steps in the task fail as n-gage has been pruned
prematurely.

I don't think we really want to be running `npm update` after we've
build the Heroku app anyway, especially now that we have lockfiles in
most projects.